reorganization of files working on server
