+++ jungfrau_gui/ui_components/file_operations/processresult_updater.py
@@ -5,36 +5,21 @@ import logging
 from datetime import datetime
 import argparse
 from pathlib import Path
-# from .. import globals
 from PySide6.QtCore import Signal, Slot, QObject
 import time
 
-# Absolute import to work in both cases: as a module and as a standalone script
-# from jungfrau_gui.ui_components.tem_controls.toolbox import config as cfg_jf
-
-# class CustomJSONEncoder(json.JSONEncoder):
-#     def default(self, obj):
-#         if isinstance(obj, (np.integer, np.int64)):
-#             return int(obj)
-#         elif isinstance(obj, (np.floating, np.float32)):
-#             return float(obj)
-#         elif isinstance(obj, np.ndarray):
-#             return obj.tolist()
-#         # Add more types as needed
-#         return super().default(obj)
-
-class ProcessedDataReceiver(QObject):
+class DataProcessingManager(QObject):
     finished = Signal()
 
-    def __init__(self, parent, host, port=3463, verbose = True, mode=0):
+    def __init__(self, parent, host, port=3467, verbose = True, mode=1):
         super().__init__()
-        self.task_name = "Processed Data Receiver"
+        self.task_name = "Processing Launcher/Receiver"
         self.parent = parent
         self.host = host
         self.port = port
         self.verbose = verbose
         if self.verbose:
-            print(f"ProcessedDataReceiver:endpoint: {self.host}:{self.port}")
+            print(f"Processing Launcher/Receiver:endpoint: {self.host}:{self.port}")
         self.trial = 0
         self.mode = mode
 
@@ -43,7 +28,7 @@ class ProcessedDataReceiver(QObject):
 
     def stop(self):
         self.trial = 0
-        logging.info("Stopping receiver...")
+        logging.info("Stopping processing client...")
 
     @Slot()
     def run(self, timeout_ms = 5000, update_interval_ms=2000, n_retry=10, verbose = True):
@@ -55,7 +40,17 @@ class ProcessedDataReceiver(QObject):
         socket.connect(f"tcp://{self.host}:{self.port}")
         self.trial = n_retry
 
-        if self.mode == 0:
+        if self.mode == 0: # emit postprocessing
+            try:
+                file_path = self.parent.visualization_panel.prev_fpath
+                socket.send_string(f"Launching the postprocess...: {file_path}")
+                result_json = socket.recv_string()
+                response = socket.recv_string()
+                if self.verbose:
+                    print(f'[dark_orange3]{self._now()} - REP: {response}[/dark_orange3]')
+            except zmq.ZMQError as e:
+                logging.error(f"Failed to launch processing on server: {e}")
+        elif self.mode == 1: # receive postprocess-result
             while self.trial > 0:
                 try:
                     socket.send_string("Results being inquired...")
@@ -76,7 +71,7 @@ class ProcessedDataReceiver(QObject):
                     self.trial -= 1
                 # finally:
                 #     # ensure the socket is closed no matter what
-        elif self.mode == 1: # load position list
+        elif self.mode == 2: # load position list
             try:
                 search_path = self.parent.visualization_panel.full_fname.text()
                 socket.send_string(f"Session-metadata being inquired...: {search_path}")
@@ -91,7 +86,7 @@ class ProcessedDataReceiver(QObject):
                     logging.info('Succeeded in loading session-metadata')
             except zmq.ZMQError as e:
                 logging.error(f"Failed to receive session-metadata request: {e}")
-        elif self.mode == 2: # send position list
+        elif self.mode == 3: # send position list
             try:
                 list_to_send = self.parent.tem_controls.tem_action.xtallist[1:]
                 list_to_send.append({'filename': self.parent.visualization_panel.full_fname.text()})
@@ -104,6 +99,21 @@ class ProcessedDataReceiver(QObject):
                     print(f'[dark_orange3]{self._now()} - REP: {response}[/dark_orange3]')
             except zmq.ZMQError as e:
                 logging.error(f"Failed to send position list to server: {e}")
+        # elif self.mode == 4: # emit re-processing with user-defined parameters
+        #     try:
+        #         list_of_dataid = self.parent.tem_controls.tem_action.xtallist[1:]
+        #         filtered_list = list_of_dataid # something filtering, e.g. by status, spots, etc.
+        #         # logging.info(filtered_list)
+        #         filtered_list.append(
+        #             {'cell': self.parent.visualization_panel.full_fname.text(),
+        #              'spacegroup': self.parent.visualization_panel.full_fname.text()})
+        #         message_json = json.dumps(filtered_list)
+        #         socket.send_string(message_json)
+        #         response = socket.recv_string()
+        #         if self.verbose:
+        #             print(f'[dark_orange3]{self._now()} - REP: {response}[/dark_orange3]')
+        #     except zmq.ZMQError as e:
+        #         logging.error(f"Failed to launch re-processing on server: {e}")
         socket.close()
         context.destroy()                
         self.finished.emit()
@@ -111,9 +121,9 @@ class ProcessedDataReceiver(QObject):
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
     parser.add_argument('-H', '--host', type=str, default="localhost", help="Host address")
-    parser.add_argument('-pt', '--port', type=int, default=3463, help="Port to bind to")
+    parser.add_argument('-pt', '--port', type=int, default=3467, help="Port to bind to")
 
     args = parser.parse_args()
 
-    receiver = ProcessedDataReceiver(host=args.host, port=args.port)
+    receiver = DataProcessingManager(host=args.host, port=args.port)
     receiver.run(timeout_ms=2000, update_interval_ms=500, n_retry=3)
